hw_4_kafka/consumer/requirements.txt
fastapi[standard]
uvicorn[standard]
SQLAlchemy
psycopg2-binary
kafka-python
